requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-bigquery==3.17.2
httpx==0.26.0
python-dotenv==1.0.0
jinja2==3.1.3
markdown==3.5.2
openai>=1.10.0
pydantic[email]==2.5.3
